some rest inerfaces added

+++ src/main/java/com/skoczo/iot_manager/dao/temp/DeviceEntityRepository.java
@@ -2,7 +2,7 @@ package com.skoczo.iot_manager.dao.temp;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface DeviceEntityRepository extends CrudRepository<DeviceEntity, Integer> {
+public interface DeviceEntityRepository extends CrudRepository<DeviceEntity, Long> {
 
 	public DeviceEntity findByDeviceId(String deviceId);
 }
