+++ src/main/java/com/skoczo/iot_manager/dao/temp/SensorEntityRepository.java
@@ -2,7 +2,7 @@ package com.skoczo.iot_manager.dao.temp;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface SensorEntityRepository extends CrudRepository<SensorEntity, Integer> {
+public interface SensorEntityRepository extends CrudRepository<SensorEntity, Long> {
 
 	public SensorEntity findBySensorId(String sensorId);
 }
